requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==5.0.1
attr==0.3.1
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.40.2
botocore==1.40.2
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
cachetools==5.5.2
Cartopy==0.25.0
certifi==2025.7.14
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560520483/work
cftime==1.6.4.post1
charset-normalizer==2.1.1
click==8.2.1
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
datamodel-code-generator==0.26.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1752827112/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
defusedxml==0.7.1
distro==1.9.0
Django==5.1.11
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
expiringdict==1.2.2
Faker==37.5.3
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.4
ijson==3.4.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
inflect==5.6.2
inflection==0.5.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1754352855579/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
isodate==0.7.2
isort==5.13.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsf==0.11.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
kiwisolver==1.4.8
label-studio==1.20.0
label-studio-sdk==1.0.18
launchdarkly-server-sdk==8.2.1
lightning-utilities==0.15.0
lockfile==0.12.2
lxml==6.0.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
netCDF4==1.7.2
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.98.0
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
ordered-set==4.0.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathspec==0.12.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663128538/work
psycopg2-binary==2.9.10
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.10.1
pyparsing==3.2.3
pyproj==3.7.1
pyRFC3339==2.0.1
pyshp==2.3.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-json-logger==2.0.4
pytorch-lightning==2.5.1
pytz==2022.7.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1754238047345/work
redis==5.2.1
referencing==0.36.2
regex==2025.7.34
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
requests-file==2.1.0
requests-mock==1.12.1
rich==14.1.0
rpds-py==0.26.0
rq==1.16.2
rsa==4.9.1
rstr==3.2.2
rules==3.4
s3transfer==0.13.1
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.34.1
shapely==2.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.3
tabulate==0.9.0
tldextract==5.3.0
toml==0.10.2
tomli==2.2.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==1.8.0
torchvision==0.16.2+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1748003301700/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
triton==2.1.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.2.0
urllib3==1.26.20
user-agents==2.2.0
wandb==0.21.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
wrapt==1.17.2
xmljson==0.2.1
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
